--- a/src/maketype.jl
+++ b/src/maketype.jl
@@ -155,8 +155,9 @@ function (==)(rn1::DiffEqBase.AbstractReactionNetwork, rn2::DiffEqBase.AbstractR
     (nr1 == nr2) || return false
     idx1 = 1:nr1
     idx2 = 1:nr2
-    issetequal(substrates.(rn1, idx1), substrates.(rn2, idx2)) || return false
-    issetequal(products.(rn1, idx1), products.(rn2, idx2)) || return false
+    issetequal(substratesymstoich.(rn1, idx1), substratesymstoich.(rn2, idx2)) || return false
+    issetequal(productsymstoich.(rn1, idx1), productsymstoich.(rn2, idx2)) || return false
+    issetequal(netsymstoich.(rn1, idx1), netsymstoich.(rn2, idx2)) || return false
     issetequal(dependants.(rn1, idx1), dependants.(rn2, idx2)) || return false
     issetequal(rateexpr.(rn1, idx1), rateexpr.(rn2, idx2)) || return false
     issetequal(ismassaction.(rn1, idx1), ismassaction.(rn2, idx2)) || return false
@@ -294,6 +295,7 @@ function addreaction!(rn::DiffEqBase.AbstractReactionNetwork, rateex::ExprValues
     substrates = ReactantStruct[ReactantStruct(p[1],p[2]) for p in subs]
     dependents = Symbol[p[1] for p in subs]
     products = ReactantStruct[ReactantStruct(p[1],p[2]) for p in prods]
+    ns = netstoich(substrates, products)
     rate_DE = mass_rate_DE(substrates, true, rateex)
     rate_SSA = mass_rate_SSA(substrates, true, rateex)   
     
@@ -319,7 +321,7 @@ function addreaction!(rn::DiffEqBase.AbstractReactionNetwork, rateex::ExprValues
         dependents = newdeps
     end
     
-    push!(rn.reactions, ReactionStruct(substrates, products, rateex, rate_DE, rate_SSA, dependents, ismassaction))
+    push!(rn.reactions, ReactionStruct(substrates, products, ns, rateex, rate_DE, rate_SSA, dependents, ismassaction))
     nothing
 end