--- a/src/reaction_network.jl
+++ b/src/reaction_network.jl
@@ -540,7 +540,6 @@ function calculate_symjac(reactions, reactants, parameters, syms)
     jacexprs = Matrix{ExprValues}(undef, nspecs, nspecs)
     fill!(jacexprs, 0)
 
-    nrxs = length(reactants)
     @inbounds for rx in reactions         
         if rx.is_pure_mass_action 
             for sub in rx.substrates