@isaacsas 2019-04-08
Some simple timings of symbolic Jacobian generation:

```
rs = @reaction_network dtype begin
           k1, 2A --> B
           k2, B --> 2A
           k3, A + B --> C
           k4, C --> A + B
           k5, 3C --> 3A; k6, 0 --> 2B
       end k1 k2 k3 k4 k5 k6
```
current release:
```
@btime DiffEqBiological.calculate_symjac(rs.f_func, rs.syms)
  914.691 μs (644 allocations: 37.81 KiB)
```
new code with mass action optimization
```
@btime DiffEqBiological.calculate_symjac(rs.reactions, rs.syms_to_ints, rs.params_to_ints, rs.syms)
  22.516 μs (99 allocations: 10.48 KiB)
```
new code without mass action optimization
```
@btime DiffEqBiological.calculate_symjac(rs.reactions, rs.syms_to_ints, rs.params_to_ints, rs.syms)
  831.295 μs (431 allocations: 20.64 KiB)
```
```
julia> rs = @reaction_network begin
           hillr(P₃,α,K,n), ∅ --> m₁
           hillr(P₁,α,K,n), ∅ --> m₂
           hillr(P₂,α,K,n), ∅ --> m₃
           (δ,γ), m₁ ↔ ∅
           (δ,γ), m₂ ↔ ∅
           (δ,γ), m₃ ↔ ∅
           β, m₁ --> m₁ + P₁
           β, m₂ --> m₂ + P₂
           β, m₃ --> m₃ + P₃
           μ, P₁ --> ∅
           μ, P₂ --> ∅
           μ, P₃ --> ∅
       end α K n δ γ β μ;
```
current release:
```
@btime DiffEqBiological.calculate_symjac(rs.f_func, rs.syms)
  2.291 ms (1416 allocations: 79.81 KiB)
```
new code with mass action optimization
```
@btime DiffEqBiological.calculate_symjac(rs.reactions, rs.syms_to_ints, rs.params_to_ints, rs.syms)
  712.815 μs (391 allocations: 26.23 KiB)
```
new code without mass action optimization
```
 @btime DiffEqBiological.calculate_symjac(rs.reactions, rs.syms_to_ints, rs.params_to_ints, rs.syms)
  1.333 ms (724 allocations: 40.16 KiB)
```