@isaacsas 2019-04-15
@ChrisRackauckas If you could look over when you have a chance that'd be great. The general idea is to try to handle mass action reactions more directly, which eliminates the SymEngine dependence for them at least. Later we can swap SymEngine for ModelingToolkit for the non-mass action reactions. Making this change meant updating most of the expression generation functions to loop through reactions, calculate terms, and them modify the corresponding `du` or `J` equation. (The generated code should be essentially the same, just the generation process should be more direct.) Updating parameter Jacobian code is still TODO.

This puts everything in place to work on sparse Jacobians, modulo the issue of needing to split sufficiently large systems into sub-methods due to Julia's ~300,000 expression list size limit.